Build: 7645.4367 Update return and method parameters
In this update, we changed the return value in the app.py, from 'prediction_id' to 'result_handler.result'. Additionally, the 'extract_labels' and 'beautify_data' methods within ResultsHandler.py had slight adjustments, where the 'self' parameter has been removed. This is done to improve the clarity of the code, and to ensure the functions are static and can be invoked without creating an instance.